worker/requirements.txt
aiohttp
face_recognition
numpy
Pillow